spec: open dropdown before clicking item
If we click on the item before the dropdown is open, we run the risk of the event listener not being available yet and the click not doing anything. Fix this by opening the dropdown before clicking the item as the user would.

refs DE-205

Change-Id: Ic481362aee7c0f964db400305ee9c68002ca4983
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/245220
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: James Butters <jbutters@instructure.com>
QA-Review: James Butters <jbutters@instructure.com>
Product-Review: Aaron Ogata <aogata@instructure.com>