set focus to first item on Canvas login forgot password form
this change is for the mobile web view of the forgot password
form. we created a tabbable interface and also added focus
to the first visible input element for ease of accessibility.

after further review our ajax callbacks were not getting
fired due to duplicate classes on a Button tag. This opened
a can of worms as it never worked before. we're getting rid
of the gross <alert> calls as that is a user and design
smell. there isn't a great way to display flash messages in
a mobile type view currently so I settled with displaying
status text on the request password button. before the
experience would be a browser alert popup that would
display undefined because the underlying code was broken.

unfortunately there's not a great testing paradigm for our
bundle code, besides the view layer, which, in the case of
ERB, is pretty gross and not very good.

fixes FOO-35
flag = none

test plan:
 - using chrome dev tools responsive view
 - navigate to /login/canvas
 - select the "Forgot Password?" link
 - verify the focus is set on the first navigable
   element that is visible. (pseudonym unique_id input)
 - you should be able to use keyboard nav to tab in and
   out of this input.
 - submitting an email or otherwise unique_id in this
   form input should *work* and process an AJAX request
   as normal after the submit event gets triggered.

Change-Id: I5a660b4d592d9bd6eee3baef71536caa85b66bfc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/245645
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>
Product-Review: August Thornton <august@instructure.com>