Don't return stale data after updating permissions
Fixes FOO-661
Refs FOO-873
flag=none

When updating a permission via the API, even though the change
is properly persisted, the stale (pre-update) value may get
returned in the API response. This is because although the
RoleOverride model clears the caches after a save, the update
of the Redis caches may happen asynchronously and thus may not
be complete by the time the response is built.

The fix is to add a new argument to RoleOverride.permissions_for
which causes it to bypass all the caches and always hit the AR
model itself. Then the role_json method in lib/api/v1 can use
that to force permissions_for to skip any cached entries.

Fixing this plus adding the spinners to the permissions buttons
should make the whole workflow of editing permissions settings
way more clear to the user.

Test plan:
* Play around with the permissions settings page
* Your changes should always be reflected properly after the
  API call completes, and should stay that way

Change-Id: I5b0fc1a242f886a4b2d2940e61d2e6d65c1d43d8
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/245905
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>