Fix duplicate tools shown at pagination breaks
fixes INTEROP-4606
flag=none

Given "Tool 1" is the last tool on a page, Canvas currently uses the
downcase'd name for the bookmark, e.g. "tool 1". When using the bookmark
(in restrict_scope()), The collation key for the tool in the database
(collation key for "Tool 1") is greater than the collation key for "tool
1", so it is included in the output for the second page as well.

I thought at first we could just not downcase the tool name in
bookmark_for().  However, for the tools page, we create bookmarkers for
the LTI 2 tools and LTI
1 tools and merge them by comparing the bookmarks for each object.

So the essential problem is: merging bookmarks needs to use same
collation technique as fetching. Canvas::ICU() appears to be _mostly_
the same collation algorithm as postgres uses, with these exceptions:
* I found a Unicode collation test and did find some _extremely_ rare
  characters that caused different orderings between Canvas::ICU() and
  public.collkey(), but I'm confident they would never be used in an LTI
  tool name and probably almost never any text of any language.
* We may soon be moving from public.collkey() to COLLATE
  public."und-u-kn-true" but I tested these too and it only differs for
  extremely rare characters.
* I did not test different locales.

However, the worst-case scenario is the bookmark merging doesn't get the
last key in the PG collation and we get duplicates again; we would never
skip (not show) any tools in the pagination.

Test plan:
- Have some tools with both lower- and upper-case letters, ideally a mix
  of LTI 1 and LTI 2 tools.
- Open ui/features/external_apps/react/lib/ExternalAppsStore.js and edit
  the PER_PAGE constant to force a page break before the end of your
  tools. Break in a place near the tools with both lower- and upper-case
  letters.
- Open the Apps listings page in account or course settings and make
  sure there are no duplicate tools listed where the page breaks (two
  different API calls are made).
- For the ultimate test try changing PER_PAGE to 1 and make sure the
  tools are still in the same order and there are none skipped or
  duplicated

Change-Id: I3169692a16f9bc0eac7bdbba94beebf73df6490d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/276660
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>