Fix explosion when updating a role (and false positive test)
Fixes FOO-896
flag=granular_permissions_wiki_pages

There was a bug involving updating roles in which the updated
role did not get the group roles added to it. This caused an
explosion in the UI when the expected group role no longer
existed in the newly-updated role. This was made worse because
the selenium test for the permissions page did not enable the
feature flag and so was blind to this error.

This fixes the bug and also makes the selenium test with the FF
on (it's soon to be removed anyway) sensitive to this case.

Test plan:
* Have a custom role
* Go into the permissions page and edit the role by updating
  its name
* That should work and not explode once the name change happens

Change-Id: Ic601772a9decf2cb1431afd90923b0ad66f72ffb
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/246077
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>