Add "Disable Keyboard Shortcuts" user feature flag
Closes FOO-871
flag=none

Create a user feature option to allow them to disable
keyboard shortcuts Canvas-wide. This'll be a lot of work
to actually do in the code, but we nee the FF here first
so that code can be conditional on it.

The way that user FFs have been fixed up is that if they're
hidden at the top level, they don't even show up at lower
levels, so creating it this way serves both as the eventual
user option (once it's set to "Allow" by default or by a
site admin), and also as the development feature flag (since
it is initially set to "Hidden")

Test plan:
* Flag shows up in site admin
* Users don't see it

Change-Id: I12390042f94139df07e6f41bd146a98314f1d6d2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/246679
Reviewed-by: Michael Ziwisky <mziwisky@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>