Remove "view longer description" from outcome
After adding a description to an outcome and attaching it to a rubric as a
criterion, there is still a "View longer description" button. This is
inconsistent with the longer descriptions on standard criterion,
which do not have that link.

closes OUT-2123
flag=none

test plan:
- View an existing rubric or create a new rubric
- Add an outcome criterion with description
- You should see outcome long description without html tags
  and media (videos)
- You shouldn't see "View longer description" button

Change-Id: I25e3c2633017eb8c49304bd3968c2642969ecc0a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/249039
Reviewed-by: Augusto Callejas <acallejas@instructure.com>
Reviewed-by: Michael Brewer-Davis <mbd@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Pat Renner <prenner@instructure.com>
Product-Review: Jody Sailor