convert clone group set JQueryUI dialog to InstUI modal
This is the only instance of this modal that gets
rendered with the prop (implicitly) for this case as
`openedFromCaution: false`. this component gets used as
a guard of caution before modifying a particular group set
in a myriad of files pertaining to groups and members.

closes FOO-910
flag = none

test plan:
 - As a Teacher navigate to a course's People page
 - Create a Group Set if not having done so already
 - Click on the hamburger menu and select |Clone Group Set|
 - Verify the title of the Group Set is populated in the
   name field and it's prepended with `(Clone)`
 - You are able to cancel changes without consequence
 - You are able to Submit the form to clone that particular
   group set
 - You are returned in your newly cloned group set

Change-Id: I58c23a1eea8e52112b640188905940f46b12fc7e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/251588
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>