close tray when LTI tool executed the success return url
When a tool redirects to the success return URL from the
assignment_index_menu placement, we need to close the tray and refresh
the current page to load the new assignments.

closes INTEROP-6264
flag=none

test-plan:
* Have an LTI tool installed in your local Canvas.
* Have a Course recorded.
* The `Assignment Index Menu` must be enabled into the `placements` for
this tool,  in case you need to change the placements, keep sure that
delayed jobs were executed to reflect this new configuration.
* Access the `Assignments` page, click into the `Assignments Settings`
button and then click into the tool;
* After the tool launch into the tray area, you can submmit the success
return url and check that the page was reloaded and the tray was closed.

Into the `lti-1.3-test-tool` we added a button called
`Call the success return URL` that will trigger the success return url.

Change-Id: If81da19060d55240f6a79252e5d9e8932470d405
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/251929
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>
QA-Review: Weston Dransfield <wdransfield@instructure.com>
Product-Review: Karl Lloyd <karl@instructure.com>