use course scales when associating rubrics
closes OUT-4046

flag=none

test-plan:
- Import a rubric into a course from both a
course and an account via `Find a Rubric` when
adding a rubric to an assignment
 > Verify that, for the course rubric, the rubric is duplicated
in the new course. This can be confirmed either by viewing
the Network tab or checking the rails console for `Rubric.last`
and checking the two rubric ids.
> Verify that, for the account rubric, a new rubric is not created
- Repeat the above steps with the `account_level_mastery_scales`
FF enabled and a rubric that has learning outcomes attached
to it
  > Create an account level mastery scale, if you dont already
have one
  > Create a course level mastery scale in the course you are
importing the rubrics into
  > Verify that the learning outcome criterion within the rubric
use the account level mastery scale for the account
rubric and the imported course mastery scale for the duplicated
course rubric

Change-Id: Icbe968f01f370ff18ed74870c8bbcbc8f323ef5d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/252895
Reviewed-by: Augusto Callejas <acallejas@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jody Sailor
QA-Review: Brian Watson <bwatson@instructure.com>