observers can only add new observees (via the UI) with pairing codes
fixes FOO-934

Why is an observer given a different "add observee" UI when SAML is the
primary auth provider?  Because of PFS-812 -- back in the day, there
were no pairing codes.  Instead, the way to add an observee was to enter
their username and password into Canvas.  This was problematic for a
particular school that used SAML logins -- users didn't _have_ a Canvas
un/pw.  So we built a different flow that forwarded the observer to the
configured SAML auth provider to login as the observee _there_ instead.
Then came pairing codes, and later came
https://gerrit.instructure.com/c/canvas-lms/+/157514 to remove the
pairing code feature flag and make it the only way for an observer to
add observees (in the UI -- the API remained and remains backwards
compatible).  Or at least that seemed to be the intention -- problem is
it actually left the SAML flow behind too, and not only that, it broke
it by changing the shape of the data that gets posted to Canvas (see the
change to `addObservee` in UserObserveesView.coffee).

This commit makes pairing codes the One True Way for an observer to add
observees (again, in the UI), regardless of the account's auth
configuration.

Note that the SAML observer registration flow -- i.e. the half of
PFS-812 that (as far as I know) hasn't been broken since the middle of
2018 -- is kept intact.

test plan:
 - follow repro steps in ticket
 - when you get to the part where you're logged in as an observer and
   you're at Account > Settings > Observing, notice the input is for a
   pairing code, not a username
 - entering a valid pairing code in that box successfully adds the
   corresponding user to your observees list.

Change-Id: I8c875b3824d76786c6d9a787de05c55dfb6dbc6d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/253344
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>