better html response detection for CSP
fixes FOO-1211
flag=none

the `request.format` is unreliable -- it can be non-html even when the
app responds with html (follow the test plan below without this change
applied to see that). use the `response.mime_type` instead -- it's a
direct indicator of what we're responding with (which is what we
actually care about) as opposed to what the request asks for.

test plan:
 - Enable CSP on your account.
   - go to /accounts/self/settings, click Feature Options, enable Content Security Policy
   - refresh, click Security tab in top nav, "Enable Content Security Policy"
 - Create a classic quiz.
 - Use HTML editor: input <iframe src="somewhere not in CSP"></iframe>
 - Save and Publish the quiz
 - Note that content is blocked.
 - Append `?format=js` to the URL and load the page
 - Note that although the content formatting is messed up, the content is still blocked
   and the CSP header is attached to the response for the page.

Change-Id: I0f5c7bdb9aa605de482dc9f944d9c7e61f723836
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/254699
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>
QA-Review: Michael Ziwisky <mziwisky@instructure.com>