Prevent video preview from overflowing into the Usage Rights UI
closes LS-1505
flag=rce_enhancements

Because (1) dealing with the size of Tabs is difficult and (2) the
<UsageRightsSelectBox> is rendered up in the Modal and not in the
ComputerPanel, we have to hard-code an upper bound to the height of
the video preview. It's not so bad, since we should have constrained the
height of the preview anyway so it fits on the screen.

test plan:
  - in the new rce, open Documents > Upload Document
  - pick a large size video
  > expect the preview not to overflow into the Usage Rights UI

Change-Id: Ia1ce716112caac430d575d998782a1b8dc3a84c1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/255379
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>