Add post policy to LO result exclude_muted_associations
Graded Assignments attached to Outcomes should show Mastery
to Students if the submission does not have grades or comments.
This is true for assignments that have a post policy of automatic.
If the post policy is set to manual, the student should not see the
Outcome result until the instructor has posted the assignment
results. By default, all assignments are set to automatic.

closes OUT-4139
flag=none

test plan
- PreReq: A published course with at least one student with the
Learning Mastery Gradebook and Student Learning Mastery
Gradebook Feature Options enabled.
- Create Assignment with 'Display Grade As' to any option other
than 'Not Graded' (Such as 'Points'), attach a Rubric/Outcome that
is not used for grading and save.
- Evaluate student with Outcome in SpeedGrader, but do not give a
score or any submission comments only save the outcome result.
- Masquerade as the student and click on Grades, you will see the
Mastery information for that assignment.
- Create second Assignment with 'Display Grade As' to any option
other than 'Not Graded' (Such as 'Points'), attach a Rubric/Outcome
that is not used for grading and save.
- Next go to Gradebook, click on 3 dot menu drop down by the name
Of the newly created assignment.  Click on Grade Posting Policy and
Set it to Manual.
- Go back to the assignment and evaluate student with Outcome in
SpeedGrader, but do not give a score or any submission comments
only save the outcome result.
- Masquerade as the student and click on Grades, you will NOT see
the Mastery information for that assignment.

Change-Id: I995690d135978110edbbe50f7970454c71044327
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/256230
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Pat Renner <prenner@instructure.com>
Product-Review: Jody Sailor
Reviewed-by: Michael Brewer-Davis <mbd@instructure.com>
Reviewed-by: Pat Renner <prenner@instructure.com>