don't set implicit course dates when shifting an import
when start/end dates are not completely specified in content imports,
canvas looks for the first and last event/due dates in the course to
help with the math. the problem is that it also uses these dates to
set the course start / conclude dates. if one of these dates is given
and the other is implied, the validation check in the migration code
is circumvented and the Course model validation fails instead, which
causes the migration to fail and some post-migration housekeeping
(such as cached due dates) does not run

I think it doesn't make sense to make the implicit start/end dates
explicit in the first place, so I prefer to take them out rather
than fix the validation.

test plan:
 - have a course with no start/end dates and one assignment
   with a due date
 - create an empty course shell to copy into. enroll a
   student in the course.
 - perform a course copy, choose to shift dates, and leave
   three dates blank, specifying only the new end date,
   and give a date that is *earlier* than the assignment's
   due date
 - the migration should succeed
 - speedgrader should be able to view the assignment in
   the destination course

fixes LS-1670

Change-Id: Ic50004fb53f91cb2d048ab47bfbcafbb410cff59
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/256404
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>